Fixing defs

+++ src/defs.hpp
@@ -6,13 +6,13 @@
 namespace OpenWars {
 	typedef uint8_t u8;
 	typedef uint16_t u16;
-	typedef uint16_t u32;
-	typedef uint16_t u64;
+	typedef uint32_t u32;
+	typedef uint64_t u64;
 	
 	typedef int8_t i8;
 	typedef int16_t i16;
-	typedef int16_t i32;
-	typedef int16_t i64;
+	typedef int32_t i32;
+	typedef int64_t i64;
 	
 	typedef float f32;
 	typedef double f64;
